+++ CHANGELOG.md
@@ -6,10 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Fixes
+### 🔧 Fixes
 - Reduced the number of CLI options sent when `IUnitTest.UnitTests` target runs and the current build does not need to collect code coverage.
 - Added CLI options to push mutation reports to stryker dashboard when possible.
-- Made `IMutationTest.MutationTests` target to run before `IPack.Pack`.
+- Made `IMutationTest.MutationTests` target to run before `IPack.Pack` does.
 
 ## [0.5.0] / 2023-07-24
 ### 🚀 New features