+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Made `IMutationTest.MutationTests` target to run before `IPack.Pack` does.
 - Added CLI option `--project` to the options output when running mutation tests
 
+
 ## [0.5.0] / 2023-07-24
 ### 🚀 New features
 - Added `ConfigName` parameter to specify the name of the configuration to use when pushing nuget packages ([#37](https://github.com/candoumbe/Pipelines/issues/37)).